move contextTypes out of the Component to fix webpack bug
