+++ src/Tree.jsx
@@ -4,6 +4,7 @@ import classNames from 'classnames';
 import warning from 'warning';
 import { polyfill } from 'react-lifecycles-compat';
 
+import { treeContextTypes } from './contextTypes';
 import {
   traverseTreeNodes, getStrictlyValue,
   convertTreeToEntities, convertDataToTree,
@@ -90,57 +91,6 @@ const getSyncProps = (props = {}, prevProps, preState) => {
   return null;
 };
 
-/**
- * Thought we still use `cloneElement` to pass `key`,
- * other props can pass with context for future refactor.
- */
-export const contextTypes = {
-  rcTree: PropTypes.shape({
-    root: PropTypes.object,
-
-    prefixCls: PropTypes.string,
-    selectable: PropTypes.bool,
-    showIcon: PropTypes.bool,
-    icon: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
-    draggable: PropTypes.bool,
-    checkable: PropTypes.oneOfType([
-      PropTypes.bool,
-      PropTypes.node,
-    ]),
-    checkStrictly: PropTypes.bool,
-    disabled: PropTypes.bool,
-    openTransitionName: PropTypes.string,
-    openAnimation: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
-
-    loadData: PropTypes.func,
-    filterTreeNode: PropTypes.func,
-    renderTreeNode: PropTypes.func,
-
-    isKeyChecked: PropTypes.func,
-
-    onNodeClick: PropTypes.func,
-    onNodeDoubleClick: PropTypes.func,
-    onNodeExpand: PropTypes.func,
-    onNodeSelect: PropTypes.func,
-    onNodeMouseEnter: PropTypes.func,
-    onNodeMouseLeave: PropTypes.func,
-    onNodeContextMenu: PropTypes.func,
-    onNodeDragStart: PropTypes.func,
-    onNodeDragEnter: PropTypes.func,
-    onNodeDragOver: PropTypes.func,
-    onNodeDragLeave: PropTypes.func,
-    onNodeDragEnd: PropTypes.func,
-    onNodeDrop: PropTypes.func,
-    onBatchNodeCheck: PropTypes.func,
-    onCheckConductFinished: PropTypes.func,
-
-    // treeData: PropTypes.array,
-    // Tree will store the entities when the treeNode refresh.
-    // User can pass the func to add more info to customize the additional info.
-    // processTreeEntity: PropTypes.func,
-  }),
-};
-
 class Tree extends React.Component {
   static propTypes = {
     prefixCls: PropTypes.string,
@@ -194,7 +144,7 @@ class Tree extends React.Component {
     openAnimation: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
   };
 
-  static childContextTypes = contextTypes;
+  static childContextTypes = treeContextTypes;
 
   static defaultProps = {
     prefixCls: 'rc-tree',
